requirements.txt
Flask
requests
markdown
pdfkit
PyPDF2
python-docx
textract
gunicorn
